% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_metab_clust_survival.R
\name{kmeans_metab_clust_surv}
\alias{kmeans_metab_clust_surv}
\title{Plot Kaplan-Meier analysis of clusters for the selected endpoint}
\usage{
kmeans_metab_clust_surv(
  kmeans_res,
  clinical_data,
  sample_id_col,
  surv_status_col,
  surv_time_col
)
}
\arguments{
\item{kmeans_res}{Data frame holding the cluster membership of each sample}

\item{clinical_data}{Data frame with survival data of the cases analyzed}

\item{sample_id_col}{Column name of sample IDs in the clinical_data table}

\item{surv_status_col}{Column name of survival status (such as overall 
survival status) in the clinical_data table}

\item{surv_time_col}{Column name of survival time (such as overall survival time) 
in the clinical_data table}
}
\value{
A Kaplan-Meier plot of the clusters for the selected endpoint
}
\description{
Creates a Kaplan-Meier (KM) plot of clusters for the selected clinical 
endpoint
}
\examples{
kmeans_metab_clust_surv(kmeans_results, clinical_table, "case_id", 
"overall_surv_status", "overall_surv_time");
}
