% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_pathway_activity.R
\name{plot_mean_pathway_activity}
\alias{plot_mean_pathway_activity}
\title{Plot mean pathway enrichment scores per cluster and pathway}
\usage{
plot_mean_pathway_activity(gsva_data, kegg_gs, kmeans_res)
}
\arguments{
\item{gsva_data}{N x M matrix with N pathways and M tumor samples}

\item{kegg_gs}{Named list with K elements of KEGG metabolic pathways gene set}

\item{kmeans_res}{Data frame holding the cluster membership of each sample}
}
\value{
A barplot of mean enrichment scores
}
\description{
Creates a barplot of mean GSVA enrichment scores per cluster and pathway 
(pathways are grouped by main KEGG metabolic category)
}
\examples{
plot_mean_pathway_activity(gsva_matrix, kegg_gs, kmeans_res);
}
