% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsva_metabolic.R
\name{run_gsva_metabolic}
\alias{run_gsva_metabolic}
\title{Gene Set Variation Analysis (GSVA) on gene expression data of tumors with 
KEGG metabolic pathways}
\usage{
run_gsva_metabolic(gene_exp_data, kcdf, kegg_gs)
}
\arguments{
\item{gene_exp_data}{N x M gene expression matrix with rows (row names) as 
genes and samples as column names}

\item{kcdf}{Character, either "Gaussian" or "Poisson" 
(see GSVA documentation)}

\item{kegg_gs}{N long named list of KEGG metabolic pathways gene set}
}
\value{
GSVA results matrix with enrichment scores of samples per KEGG 
metabolic pathway
}
\description{
Performs Gene Set Variation Analysis (GSVA) on gene expression matrix of 
tumor samples using KEGG metabolic pathways gene set collection
}
\examples{
run_gsva_metabolic(gene_exp_matrix, "Gaussian", kegg_gs);
}
