% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_progeny.R
\name{run_progeny}
\alias{run_progeny}
\title{Runs PROGENy on the input gene expression data}
\usage{
run_progeny(gene_exp_data, kmeans_res, vst = FALSE, top = 100)
}
\arguments{
\item{gene_exp_data}{N x M gene expression matrix genes as 
row names and samples as column names}

\item{kmeans_res}{Data frame holding the cluster membership of each sample}

\item{vst}{Logical, whether to do VST on the gene expression data or not
(default: FALSE)}

\item{top}{Numeric, he top n genes for generating the model matrix according 
to significance (p-value) (see PROGENy documentation for details)}
}
\value{
A data frame with PROGENy pathway activity scores per sample, and
a boxplot showing the differential pathway testing results between clusters
}
\description{
Runs PROGENy on the input gene expression data and compares PROGENy pathway
activtiy between the identified clusters. Wilcoxon Rank Sum test is used if 
number of clusters = 2, if number of clusters > 2 then Kruskal-Wallis Rank 
Sum test is used.
}
\examples{
run_progeny(gene_exp_data, kmeans_res, vst = FALSE, top = 100);
}
